# Conflicts:
#	build.gradle

+++ src/main/java/io/th0rgal/oraxen/mechanics/provided/farming/bottledexp/BottledExpMechanic.java
@@ -18,9 +18,9 @@ public class BottledExpMechanic extends Mechanic {
             return level * level + 6 * level;
 
         if (level <= 30)
-            return (int) (2.5 * (double) level * (double) level - 40.5 * (double) level + 360.0);
+            return (int) (2.5 * level * level - 40.5 * level + 360.0);
 
-        return (int) (4.5 * (double) level * (double) level - 162.5 * (double) level + 2220.0);
+        return (int) (4.5 * level * level - 162.5 * level + 2220.0);
     }
 
     public int getBottleEquivalent(int level, float xp) {
